Fixed grammar and checked clarity
I fixed the grammar and checked clarity but I think mostly what I found may need to be fixed by someone else other than me. It's mostly just clarity and understanding what the true purpose of this page is. Is this meant to give instructions for people who want to run and build k-means or is it something else?